requirements.txt
asgiref==3.7.2
Django==3.2.23
djangorestframework==3.12.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
